+++ CHANGELOG.md
@@ -29,6 +29,7 @@ v1.0.0 (in development)
 - Add a `RepositoryPage` class for representing the return value of
   `parse_repo_links()` (now called `RepositoryPage.from_html()`)
 - Renamed `DistributionPackage.from_pep691_details()` to `from_json_data()`
+- `PyPISimple.stream_project_names()` now accepts JSON responses
 
 v0.10.0 (2022-06-30)
 --------------------